requirements.txt
boto3~=1.35.34
python-dotenv~=1.0.1
mysql-connector-python
faker